+++ pages/signin/index.tsx
@@ -1,4 +1,4 @@
-import SignInForm from "@/src/sign/signin/SignInForm";
+import SignInForm from "@/src/sign/Components/signin/SignInForm";
 
 const SignIn = () => {
   return(
